+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Add command `nf-class modules create-from-class` [#1](https://github.com/mirpedrol/nf-class/pull/1)
 - Add command `nf-class subworkflows expand-class` [#2](https://github.com/mirpedrol/nf-class/pull/2)
 - Add pytests [#5](https://github.com/mirpedrol/nf-class/pull/5)
+- Handle composed modules when expanding a subworkflow [#6](https://github.com/mirpedrol/nf-class/pull/6)